problems/src/main/java/com/github/ajanthan/problems/p160/Solution.java
package com.github.ajanthan.problems.p160;

import com.github.ajanthan.problems.ListNode;

public class Solution {
    public ListNode getIntersectionNodeV0(ListNode headA, ListNode headB) {
        int lengthA = getLength(headA);
        int lengthB = getLength(headB);
        if (lengthA > lengthB) {
            for (int i = 1; i <= lengthA - lengthB; i++) {
                headA = headA.next;
            }
        } else if (lengthA < lengthB) {
            for (int i = 1; i <= lengthB - lengthA; i++) {
                headB = headB.next;
            }
        }
        while (headA != null && headB != null) {
            if (headA == headB) {
                return headA;
            }
            headA = headA.next;
            headB = headB.next;
        }
        return null;
    }

    private int getLength(ListNode head) {
        int length = 0;
        while (head != null) {
            length++;
            head = head.next;
        }
        return length;
    }

    public ListNode getIntersectionNodeV1(ListNode headA, ListNode headB) {
        ListNode pA = headA, pB = headB;
        while (pA != pB) {
            pA = pA == null ? headB : pA.next;
            pB = pB == null ? headA : pB.next;
        }
        return pA;
    }
}
